+++ flycheck-ert.el
@@ -385,6 +385,9 @@ resource directory."
       (ert-skip (format "%S missing" mode)))
     (flycheck-ert-with-resource-buffer resource-file
       (funcall mode)
+      ;; Load safe file-local variables because some tests depend on them
+      (let ((enable-local-variables :safe))
+        (hack-local-variables))
       ;; Configure config file locating for unit tests
       (let ((process-hook-called 0))
         (add-hook 'flycheck-process-error-functions